pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ai-job-market-dashboard"
version = "1.0.0"
description = "AI-Driven Job Market Intelligence Dashboard ETL Pipeline"
requires-python = ">=3.11"
readme = "README.md"
dependencies = [
    "pandas>=2.0.0",
    "pyarrow>=12.0.0",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
    "sentence-transformers>=2.2.0",
    "scikit-learn>=1.3.0",
    "prophet>=1.1.4",
    "numpy>=1.24.0",
    "geopy>=2.3.0",
    "openai>=1.0.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
]

[tool.setuptools]
# Mark as a non-package project (scripts only)
packages = []
py-modules = []

[tool.setuptools.package-data]
"*" = ["*.csv", "*.yml", "*.yaml"]

[tool.black]
line-length = 100
target-version = ['py311']

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true

